Extract event handlers to separate module + implement IP freeing (OMS-297)
